requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
psutil>=5.9.0
aiohttp-session>=2.12.0
cryptography>=41.0.0
wavelink>=3.0.0
